==> Feature-Engineering/2. NER on Tags.ipynb <==
#Installing packages
!python -m spacy download en_core_web_lg

# Importing libraries
from keras.utils.np_utils import to_categorical

import pandas as pd
import numpy as np

import re

import sklearn
from sklearn.feature_extraction.text import TfidfVectorizer
from itertools import combinations
from sklearn.cluster import KMeans

from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA

import tensorflow as tf
import tensorflow.compat.v1 as tf
import tensorflow_hub as hub
import numpy as np

import matplotlib.pyplot as plt
import nltk

nltk.download('averaged_perceptron_tagger')
nltk.download('maxent_ne_chunker')
nltk.download('words')
nltk.download('stopwords')

nltk.download('punkt')

import spacy 
from spacy import displacy

import random

# from rake_nltk import Rake

from numpy import array
from tensorflow.keras.preprocessing.text import one_hot
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Flatten,Embedding,Dense

import os
import tensorflow as tf
import tensorflow_datasets as tfds
from tensorboard.plugins import projector

from gensim.scripts.glove2word2vec import glove2word2vec

# import en_core_web_lg
# nlp = en_core_web_lg.load()

from nltk.tokenize import word_tokenize

import gensim
from scipy import spatial

import csv
import tensorflow as tf

# import transformers
# transformers.logging.set_verbosity(transformers.logging.CRITICAL)

# from sentence_transformers import SentenceTransformer, util
# import nmslib

# distilbert = SentenceTransformer('distilbert-base-uncased')

import multiprocessing
import math

import time
from gensim.models import KeyedVectors
from gensim.models.fasttext import FastText
import pandas as pd
import gensim
gensim.__version__

#Importing the cleaned dataset
movtag2 = pd.read_pickle("/content/drive/MyDrive/MScA - UChicago/Assignments/Quarter 3 - Spring 2022/Capstone Project/Final datasets/1_cleaned_tags.pkl")

#extracting person, organisation, location and norp from the tags
%%time

person = []
org = []
gpe = []
norp = []
tags = []

for i in movtag2['combined_tags_nodup']:
  # random.shuffle(i)
  txt = ' '.join(i)
  doc = nlp(txt)

  entities = []
  labels = []


  for ent in doc.ents:
      entities.append(ent.text)
      labels.append(ent.label_)
      # position_start.append(ent.ORG)
      # position_end.append(ent.end_char)

      df_text = pd.DataFrame({'Entities':entities,'Labels':labels})
  
  lst_people = list(df_text[df_text.Labels == "PERSON"]['Entities'])
  lst_org = list(df_text[df_text.Labels == "ORG"]['Entities'])
  lst_gpe = list(df_text[(df_text.Labels == "GPE") | (df_text.Labels == "LOC")]['Entities'])
  lst_norp = list(df_text[df_text.Labels == "NORP"]['Entities'])

  lst_tags = [elem for elem in i if elem not in lst_people + lst_org + lst_gpe + lst_norp]

  person.append(lst_people)
  org.append(lst_org)
  gpe.append(lst_gpe)
  norp.append(lst_norp)
  tags.append(lst_tags)

movtag2['tag_people'] = person
movtag2['tag_org'] = org
movtag2['tag_norp'] = norp
movtag2['tag_location'] = gpe
movtag2['tags'] = tags

movtag2.head(5)

#cleaning the text further: removing words which are not present in the english dictionary
words = set(nltk.corpus.words.words())
clean_tags = []
for k in movtag2['tags']:
  lst = []
  for i in [w for w in k]:
    j = i.split(" ")
    lst.append([l for l in j if l in words])

  lst = [' '.join(w) for w in lst]
  lst = [w for w in lst if w != '']

  clean_tags.append(lst)

movtag2['tags'] = clean_tags

movtag2.drop(['genres', 'combined_tags', 'genres_nodup'], axis=1, inplace = True)

movtag2.head(2)

movtag2.to_pickle("/content/drive/MyDrive/MScA - UChicago/Assignments/Quarter 3 - Spring 2022/Capstone Project/Final datasets/2_ner_tags.pkl")

from google.colab import drive
drive.mount('/content/drive')

data = pd.read_pickle("/content/drive/MyDrive/MScA - UChicago/Assignments/Quarter 3 - Spring 2022/Capstone Project/Final datasets/2_ner_tags.pkl")

data[9:10]
