==> Feature-Engineering/5. Glove Embeddings.ipynb <==
# Importing libraries
from keras.utils.np_utils import to_categorical

import pandas as pd
import numpy as np

import re

import sklearn
from sklearn.feature_extraction.text import TfidfVectorizer
from itertools import combinations
from sklearn.cluster import KMeans

from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA

import tensorflow as tf
import tensorflow.compat.v1 as tf
import tensorflow_hub as hub
import numpy as np

import matplotlib.pyplot as plt
import nltk

nltk.download('averaged_perceptron_tagger')
nltk.download('maxent_ne_chunker')
nltk.download('words')
nltk.download('stopwords')

nltk.download('punkt')

import spacy 
from spacy import displacy

import random

# from rake_nltk import Rake

from numpy import array
from tensorflow.keras.preprocessing.text import one_hot
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Flatten,Embedding,Dense

import os
import tensorflow as tf
import tensorflow_datasets as tfds
from tensorboard.plugins import projector

from gensim.scripts.glove2word2vec import glove2word2vec

# import en_core_web_lg
# nlp = en_core_web_lg.load()

from nltk.tokenize import word_tokenize

import gensim
from scipy import spatial

import csv
import tensorflow as tf

# import transformers
# transformers.logging.set_verbosity(transformers.logging.CRITICAL)

# from sentence_transformers import SentenceTransformer, util
# import nmslib

# distilbert = SentenceTransformer('distilbert-base-uncased')

import multiprocessing
import math

import time
from gensim.models import KeyedVectors
from gensim.models.fasttext import FastText
import pandas as pd
import gensim
gensim.__version__

from keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences
from nltk.corpus import words

from keras.models import *
from keras.layers import *
from keras.callbacks import *

#importing dataset
data = pd.read_csv("/content/drive/MyDrive/MScA - UChicago/Assignments/Quarter 3 - Spring 2022/Capstone Project/ml-25m/movie_all_info.csv")
movtag4 = pd.read_pickle("/content/drive/MyDrive/MScA - UChicago/Assignments/Quarter 3 - Spring 2022/Capstone Project/Final datasets/2_ner_tags.pkl")

movtag4.head(2)

embeddings_index = dict()
f = open('/content/drive/MyDrive/MScA - UChicago/Assignments/Quarter 3 - Spring 2022/Capstone Project/glove.6B.100d.txt')

for line in f:
    values = line.split()
    word = values[0]
    coefs = np.asarray(values[1:], dtype='float32')
    embeddings_index[word] = coefs

f.close()
print('Loaded %s word vectors.' % len(embeddings_index))

# from keras.preprocessing.text import Tokenizer
# from tensorflow.keras.preprocessing.sequence import pad_sequences
# from nltk.corpus import words

def Glove_embeddings(data, ratings_dataset, colname):
  #creating the vocabulary
  tokenizer = Tokenizer()
  data[colname] = data[colname].apply(lambda x: ' '.join(x))
  x_tr = data[colname]
  tokenizer.fit_on_texts(x_tr)
  x_tr_seq  = tokenizer.texts_to_sequences(x_tr)
  x_tr_seq  = pad_sequences(x_tr_seq, maxlen=100)
  size_of_vocabulary=len(tokenizer.word_index) + 1

  #Generating the embedding Matrix
  embedding_matrix = np.zeros((size_of_vocabulary, 100))

  for word, i in tokenizer.word_index.items():
    embedding_vector = embeddings_index.get(word)
    if embedding_vector is not None:
      embedding_matrix[i] = embedding_vector
  
  #Building the sequential model
  model=Sequential()
  model.add(Embedding(size_of_vocabulary,100,weights=[embedding_matrix],input_length=100,trainable=False)) 
  model.add(LSTM(128,return_sequences=True,dropout=0.2))
  model.add(GlobalMaxPooling1D())
  model.add(Dense(64,activation='relu')) 
  model.add(Dense(10,activation='sigmoid')) 
  model.compile(optimizer='adam', loss='categorical_crossentropy',metrics=["acc"]) 
  es = EarlyStopping(monitor='val_loss', mode='min', verbose=1,patience=3)  
  mc=ModelCheckpoint('best_model.h5', monitor='val_acc', mode='max', save_best_only=True,verbose=1)  
  print(model.summary())

  #Creating labels to train the NN
  y_tr = ratings_dataset['averageRating_imdb']
  y_tr = np.array(y_tr)
  y_tr[np.where(np.isnan(y_tr))] = 0
  y_tr = to_categorical(y_tr)

  #Fitting the model
  history = model.fit(np.array(x_tr_seq),np.array(y_tr),batch_size=128,epochs=1)

  #Extracting the GloVe Embeddings and appending to the main dataframe.
  embeddings = model.layers[0].get_weights()[0]
  embeddings.shape
  data[colname] = data[colname].apply(lambda x: word_tokenize(x))
  glove_embed = []
  count = 0
  for i in data[colname]:
    lst = []
    count = count + 1
    for j in range(len(i)):
      try:
        lst.append(embeddings[tokenizer.word_index[i[j]]])
      except:
        pass
    emb = np.array(lst)
    glove_embed.append(np.sum(emb, axis = 0)/len(emb))

  data['glove_' + colname] = glove_embed

# Creating GloVe Embeddings for tag_people
Glove_embeddings(movtag4, data, 'tag_people')

# Creating GloVe Embeddings for tag_org
Glove_embeddings(movtag4, data, 'tag_org')

# Creating GloVe Embeddings for tag_norp
Glove_embeddings(movtag4, data, 'tag_norp')

# Creating GloVe Embeddings for tag_location
Glove_embeddings(movtag4, data, 'tag_location')

# Creating GloVe Embeddings for tags
Glove_embeddings(movtag4, data, 'tags')

movtag4.to_pickle("/content/drive/MyDrive/MScA - UChicago/Assignments/Quarter 3 - Spring 2022/Capstone Project/Final datasets/5_Glove_tags.pkl")

movtag4.head()
